refactor(web): authenticateにUI専用public APIを設けstorybookの巻き込みを解消
storybook の shell（AppSidebar/AppHeader）が fat barrel を import すると、
バレルが再export する useSession（swr）や検証ロジック等の .ts が storybook の
バンドル・カバレッジ母数に巻き込まれていた。UI 専用の public エントリ（ui.ts）を
設け、shell はそこから LogoutButton を取得することで、storybook が読み込む
ファイルを .tsx に限定する（client/server と同じ「互いに素なロード」を回復）。

これにより暫定対応を全廃する:
- coverage の storybook 向け .ts 除外
- optimizeDeps の swr / pino
