+++ application/packages/web/src/client/features/authenticate/index.ts
@@ -1,5 +1,4 @@
-export { AuthenticateForm } from './components/authenticate-form'
-export { default as LogoutButton } from './components/logout-button'
+export { AuthenticateForm, LogoutButton } from './ui'
 export {
   AUTH_ERROR_MESSAGES,
   resolveLoginErrorMessage,
